--- src/main/java/racingcargame/RacingCarName.java
@@ -1,13 +1,0 @@
-package racingcargame;
-
-public class RacingCarName {
-    private String name;
-
-    RacingCarName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
